handson/shin-bashoutter-1/requirements.txt
aws-cdk.aws-s3==1.100.0
aws-cdk.aws-s3-deployment==1.100.0
aws-cdk.aws-dynamodb==1.100.0
aws-cdk.aws-lambda==1.100.0
aws-cdk.aws-apigateway==1.100.0
aws-cdk.aws-cloudfront==1.100.0
aws-cdk.aws-cloudfront-origins==1.100.0
aws-cdk.aws-route53==1.100.0
aws-cdk.aws-route53-targets==1.100.0
boto3
requests
httpie
